move some queries to mandataris folder
